Fix conversations (#2556)
* fix conversations

* cleanup ConversationsRemoteMediator

* update conversation timestamps regularly

* improve loadStateListener

* add db migration

* make deleting from conversation db suspending

* reorganize code in ConversationsFragment

* delete NetworkStateViewHolder

* cleanup imports

* add 38.json

* honor fabHide setting in ConversationsFragment

* set page size to 30

+++ app/src/main/java/com/keylesspalace/tusky/components/conversation/ConversationsRemoteMediator.kt
@@ -4,8 +4,11 @@ import androidx.paging.ExperimentalPagingApi
 import androidx.paging.LoadType
 import androidx.paging.PagingState
 import androidx.paging.RemoteMediator
+import androidx.room.withTransaction
 import com.keylesspalace.tusky.db.AppDatabase
 import com.keylesspalace.tusky.network.MastodonApi
+import com.keylesspalace.tusky.util.HttpHeaderLink
+import retrofit2.HttpException
 
 @OptIn(ExperimentalPagingApi::class)
 class ConversationsRemoteMediator(
@@ -14,38 +17,53 @@ class ConversationsRemoteMediator(
     private val db: AppDatabase
 ) : RemoteMediator<Int, ConversationEntity>() {
 
+    private var nextKey: String? = null
+
+    private var order: Int = 0
+
     override suspend fun load(
         loadType: LoadType,
         state: PagingState<Int, ConversationEntity>
     ): MediatorResult {
 
+        if (loadType == LoadType.PREPEND) {
+            return MediatorResult.Success(endOfPaginationReached = true)
+        }
+
+        if (loadType == LoadType.REFRESH) {
+            nextKey = null
+            order = 0
+        }
+
         try {
-            val conversationsResult = when (loadType) {
-                LoadType.REFRESH -> {
-                    api.getConversations(limit = state.config.initialLoadSize)
-                }
-                LoadType.PREPEND -> {
-                    return MediatorResult.Success(endOfPaginationReached = true)
-                }
-                LoadType.APPEND -> {
-                    val maxId = state.pages.findLast { it.data.isNotEmpty() }?.data?.lastOrNull()?.lastStatus?.id
-                    api.getConversations(maxId = maxId, limit = state.config.pageSize)
-                }
+            val conversationsResponse = api.getConversations(maxId = nextKey, limit = state.config.pageSize)
+
+            val conversations = conversationsResponse.body()
+            if (!conversationsResponse.isSuccessful || conversations == null) {
+                return MediatorResult.Error(HttpException(conversationsResponse))
             }
 
-            if (loadType == LoadType.REFRESH) {
-                db.conversationDao().deleteForAccount(accountId)
+            db.withTransaction {
+
+                if (loadType == LoadType.REFRESH) {
+                    db.conversationDao().deleteForAccount(accountId)
+                }
+
+                val linkHeader = conversationsResponse.headers()["Link"]
+                val links = HttpHeaderLink.parse(linkHeader)
+                nextKey = HttpHeaderLink.findByRelationType(links, "next")?.uri?.getQueryParameter("max_id")
+
+                db.conversationDao().insert(
+                    conversations
+                        .filterNot { it.lastStatus == null }
+                        .map {
+                            it.toEntity(accountId, order++)
+                        }
+                )
             }
-            db.conversationDao().insert(
-                conversationsResult
-                    .filterNot { it.lastStatus == null }
-                    .map { it.toEntity(accountId) }
-            )
-            return MediatorResult.Success(endOfPaginationReached = conversationsResult.isEmpty())
+            return MediatorResult.Success(endOfPaginationReached = nextKey == null)
         } catch (e: Exception) {
             return MediatorResult.Error(e)
         }
     }
-
-    override suspend fun initialize() = InitializeAction.LAUNCH_INITIAL_REFRESH
 }
